chore: extract symbols from dashboard index (move-only)
